% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GKTfunctions.R
\name{countOutcome}
\alias{countOutcome}
\title{countOutcome}
\usage{
countOutcome(data, index, response)
}
\arguments{
\item{response}{}
}
\value{
the vector of the lagged cumulative sum.
}
\description{
Compute the prior sum of the response appearing in the outcome column for the index
}
