% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{largerawsample}
\alias{largerawsample}
\title{Raw Memphis DataShop sample}
\format{
A data frame with the original exported DataShop columns.
}
\source{
\url{https://pslcdatashop.web.cmu.edu/Export?datasetId=5513}
}
\usage{
largerawsample
}
\description{
A raw trial-level sample from the Memphis DataShop export used in examples.
}
\keyword{datasets}
