% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LKT-package.R
\docType{package}
\name{LKT-package}
\alias{LKT-package}
\title{LKT: Logistic Knowledge Tracing}
\description{
The LKT package builds logistic knowledge tracing models from educational
trial data. Its primary workflow is to compute learning-history features for
one or more components and fit a logistic model that predicts the next trial
outcome.
}
\author{
\strong{Maintainer}: Philip I. Pavlik Jr. \email{imrryr@gmail.com} (\href{https://orcid.org/0000-0001-6467-9452}{ORCID}) [contributor]

Authors:
\itemize{
  \item Philip I. Pavlik Jr. \email{imrryr@gmail.com} (\href{https://orcid.org/0000-0001-6467-9452}{ORCID}) [contributor]
  \item Luke G. Eglington \email{luke.eglington.mail@gmail.com} (\href{https://orcid.org/0000-0002-8432-9203}{ORCID}) [contributor]
}

}
\keyword{internal}
