% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LKTfunctions.R
\name{LKT_cv}
\alias{LKT_cv}
\title{LKT_cv}
\usage{
LKT_cv(
  componentl,
  featl,
  offsetl = NA,
  fixedl,
  seedl = NA,
  elastictest = FALSE,
  outputFilePath,
  val,
  cvSwitch = NULL,
  makeFolds = NULL,
  dualfit = FALSE,
  interc = FALSE
)
}
\description{
student-stratified 5 fold split half cross-validation
}
