% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LKTfunctions.R
\name{rlvl}
\alias{rlvl}
\title{rlvl}
\usage{
rlvl(dat)
}
\arguments{
\item{dat}{copy of main data frame.}
}
\value{
dataframe with 1 as first value in CF..ansbin. column
}
\description{
sorts dataframe so first student is one with an initial CF..ansbin.==1. Hack to deal with liblinear reference levels
}
