% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{samplelkt}
\alias{samplelkt}
\title{Sample practice trial sequences}
\format{
A data frame with 2074 rows and many variables:
\describe{
\item{Anon.Student.Id}{unique identifier for each student}
\item{Duration..sec.}{response duration in seconds}
\item{KC..Default.}{default knowledge component label}
\item{Outcome}{trial outcome label}
}
}
\source{
\url{https://pslcdatashop.web.cmu.edu/DatasetInfo?datasetId=5508}
}
\usage{
samplelkt
}
\description{
A small trial-level dataset from a practice memory experiment.
}
\keyword{datasets}
