% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GKTfunctions.R
\name{computeSpacingPredictors}
\alias{computeSpacingPredictors}
\title{computeSpacingPredictors}
\usage{
computeSpacingPredictors(data, KCs)
}
\arguments{
\item{data}{is a dataset with Anon.Student.Id and CF..ansbin.}

\item{KCs}{are the components for which spaced features will be specified in gkt}
}
\value{
data which is the same frame with the added spacing relevant columns.
}
\description{
Compute repetition spacing time based features from input data CF..Time. and/or CF..reltime.

which will be automatically computed from Duration..sec. if not present themselves.
}
