% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LKTfunctions.R
\name{countOutcomeold}
\alias{countOutcomeold}
\title{countOutcome}
\usage{
countOutcomeold(data, index, response)
}
\arguments{
\item{data}{the dataset to compute an outcome vector for}

\item{index}{the subsets to count over}

\item{response}{the actually response value being counted}
}
\value{
the vector of the lagged cumulative sum.
}
\description{
Compute the prior sum of the response appearing in the outcome column for the index
}
