% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{largerawsample}
\alias{largerawsample}
\title{Trial sequences for practice participants.}
\format{
A data frame please see the DataShop for more info.
\describe{
It has many columns.
}
}
\source{
\url{https://datashop.memphis.edu/Export?datasetId=1465}
}
\usage{
largerawsample
}
\description{
A dataset containing a raw sample from the Memphis Datashop.
}
\keyword{datasets}
