% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LKTfunctions.R
\name{LASSOLKTData}
\alias{LASSOLKTData}
\title{LASSOLKTData}
\usage{
LASSOLKTData(
  data,
  gridpars,
  allcomponents,
  allfeatures,
  preset = NA,
  presetint = T,
  specialcomponents = c(),
  specialfeatures = c(),
  specialpars = c(),
  removefeat = c(),
  removecomp = c()
)
}
\arguments{
\item{data}{is a dataset with Anon.Student.Id and CF..ansbin.}

\item{gridpars}{a vector of parameters to create each feature at}

\item{allcomponents}{is search space for LKT components}

\item{allfeatures}{is search space for LKT features}

\item{preset}{One of "static","AFM","PFA","advanced","AFMLLTM","PFALLTM","advancedLLTM"}

\item{presetint}{should the intercepts be included for preset components}

\item{specialcomponents}{add special components (not crossed with features, only paired with special features 1 for 1)}

\item{specialfeatures}{features for each special component (not crossed during search)}

\item{specialpars}{parameters for the special features (if needed)}

\item{removefeat}{Character Vector | Excludes specified features from the test list.}

\item{removecomp}{Character Vector | Excludes specified components from the test list.}
}
\value{
data which is the same frame with the added spacing relevant columns.

list of values "tracetable" and "currentfit"
}
\description{
Forward and backwards stepwise search for a set of features and components

with tracking of nonlinear parameters.
}
