% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LKTfunctions.R
\name{LASSOLKTModel}
\alias{LASSOLKTModel}
\title{LASSOLKTModel}
\usage{
LASSOLKTModel(
  data,
  gridpars,
  allcomponents,
  preset = NA,
  presetint = T,
  allfeatures,
  specialcomponents = c(),
  specialfeatures = c(),
  specialpars = c(),
  target_n,
  removefeat = c(),
  removecomp = c(),
  test_fold = 1,
  interc = F
)
}
\arguments{
\item{data}{is a dataset with Anon.Student.Id and CF..ansbin.}

\item{gridpars}{a vector of parameters to create each feature at}

\item{allcomponents}{is search space for LKT components}

\item{preset}{One of "static","AFM","PFA","advanced","AFMLLTM","PFALLTM","advancedLLTM"}

\item{presetint}{should the intercepts be included for preset components}

\item{allfeatures}{is search space for LKT features}

\item{specialcomponents}{add special components (not crossed with features, only paired with special features 1 for 1)}

\item{specialfeatures}{features for each special component (not crossed during search)}

\item{specialpars}{parameters for the special features (if needed)}

\item{target_n}{chosen number of features in model}

\item{removefeat}{Character Vector | Excludes specified features from the test list.}

\item{removecomp}{Character Vector | Excludes specified components from the test list.}

\item{test_fold}{the fold that the chosen LASSO model will be tested on}

\item{interc}{TRUE or FALSE, include a global intercept.}
}
\value{
list of matrices and values "train_x","train_y","test_x","test_y","fit","target_auc","target_rmse","n_features","auc_lambda","rmse_lambda","BIC_lambda","target_idx", "preds"
}
\description{
runs LASSO search on the data
}
