% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LKTfunctions.R
\name{LASSOLKTModel}
\alias{LASSOLKTModel}
\title{LASSOLKTModel}
\usage{
LASSOLKTModel(
  data,
  gridpars,
  allcomponents,
  preset = NA,
  presetint = T,
  allfeatures,
  specialcomponents = c(),
  specialfeatures = c(),
  specialpars = c(),
  target_n
)
}
\arguments{
\item{data}{is a dataset with Anon.Student.Id and CF..ansbin.}

\item{gridpars}{a vector of parameters to create each feature at}

\item{allcomponents}{is search space for LKT components}

\item{preset}{One of "static","AFM","PFA","advanced","AFMLLTM","PFALLTM","advancedLLTM"}

\item{presetint}{should the intercepts be included for preset components}

\item{allfeatures}{is search space for LKT features}

\item{specialcomponents}{add special components (not crossed with features, only paired with special features 1 for 1)}

\item{specialfeatures}{features for each special component (not crossed during search)}

\item{specialpars}{parameters for the special features (if needed)}

\item{target_n}{chosen number of features in model}
}
\value{
list of values "dropped 1se", "retained 1se","target features","target dropped","target pseudo R2","best pseudo R2","target mod rmse","target mod auc", and "target_mod_bic"
}
\description{
runs LASSO search on the data
}
