% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LKTfunctions.R
\name{LKT_HDI}
\alias{LKT_HDI}
\title{LKT_HDI}
\usage{
LKT_HDI(
  dat,
  n_boot,
  n_students,
  componentsv,
  featuresv,
  interactsv = NA,
  fixedparsv,
  get_hdi = TRUE,
  cred_mass = 0.95
)
}
\arguments{
\item{dat}{Dataframe}

\item{n_boot}{Number of subsamples to fit}

\item{n_students}{Number of students per subsample}

\item{get_hdi}{boolean to decide if generating HDI per coefficient}

\item{cred_mass}{credibility mass parameter to decide width of HDI}

\item{components}{components in model}

\item{features}{features in model}

\item{interacts}{interacts in model}

\item{fixedpars}{fixed pars in model}
}
\value{
list of values "par_reps", "mod_full", "coef_hdi"
}
\description{
Bootstrap credibility intervals to aid in interpreting coefficients.
}
