% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{samplelkt}
\alias{samplelkt}
\title{Trial sequences for practice participants.}
\format{
A data frame with 2074 rows and many variables:
\describe{
\item{Anon.Student.Id}{unique identifier for each student}
\item{Duration..sec.}{unique identifier for each student}
\item{KC..Default.}{unique identifier for each student}
\item{Outcome}{unique identifier for each student}
...
}
}
\source{
\url{http://www.datashop.memphis.edu/}
}
\usage{
samplelkt
}
\description{
A dataset containing a small sample of participants
in a memory experiment.
}
\keyword{datasets}
